fix: 캘린더 페이지 SSR 비활성화로 인증 오류 임시 대응
- SSR 경로에서 발생하는 인증/재발급 오류로 인한 렌더 실패를 임시 우회
